% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell.R
\name{cell}
\alias{cell}
\title{Cell server logic}
\usage{
cell(input, output, session, cell_id, cell.session)
}
\arguments{
\item{id}{id of the notebook}
}
\value{
A shinyApp object which run in browser (see https://shiny.rstudio.com/reference/shiny/latest/shinyApp.html for more information).
}
\description{
If \code{id=NULL} the function runs an new empty notebook, else the notebook given as argument is run.
}
\examples{
cellUI(id = 3)

}
