% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notebook.R
\name{notebookServer}
\alias{notebookServer}
\title{Notebook Server}
\usage{
notebookServer(input, output)
}
\arguments{
\item{id}{id of the notebook}
}
\value{
A shinyApp object which run in browser (see https://shiny.rstudio.com/reference/shiny/latest/shinyApp.html for more information).
}
\description{
If \code{id=NULL} the function runs an new empty notebook, else the notebook given as argument is run.
}
\examples{
runNotebook()

}
