% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NotebookSession.R
\docType{class}
\name{SessionVariables-class}
\alias{SessionVariables-class}
\alias{NotebookSession}
\title{SessionVariables object}
\description{
Container of the session variables.
}

\examples{
SessionVariables()

}
